refactor: drop New prefix from create project form
